Move NgramDocument type to mei_parsing_types.py
